Moved `EventQueue` to `event_queue.rs`
